--- src/pages/HomePage/Box/index.ts
@@ -1,2 +1,0 @@
-export { default as BoxCategory } from './BoxCategory/BoxCategory';
-export { default as BoxTodayTodo } from './BoxTodayTodo/BoxTodayTodo';
