--- src/pages/HomePage/Button/index.ts
@@ -1,2 +1,0 @@
-export { default as ButtonMoreFriends } from './ButtonMoreFriends/ButtonMoreFriends';
-export { default as ButtonUserProfile } from './ButtonUserProfile/ButtonUserProfile';
