+++ src/pages/HomePage/HomePage.tsx
@@ -29,10 +29,13 @@ import LargePlusIcon from '@/shared/assets/svgs/large_plus.svg?react';
 
 import { ROUTES_CONFIG } from '@/router/routesConfig';
 
-import { BoxCategory, BoxTodayTodo } from './Box';
-import { ButtonMoreFriends, ButtonUserProfile } from './Button';
+import BoxCategory from './BoxCategory/BoxCategory';
+import BoxTodayTodo from './BoxTodayTodo/BoxTodayTodo';
+import ButtonMoreFriends from './ButtonMoreFriends/ButtonMoreFriends';
+import ButtonUserProfile from './ButtonUserProfile/ButtonUserProfile';
 import DatePicker from './DatePicker/DatePicker';
-import { ModalContentsCategory, ModalContentsFriends } from './ModalContents';
+import ModalContentsCategory from './ModalContentsCategory/ModalContentsCategory';
+import ModalContentsFriends from './ModalContentsFriends/ModalContentsFriends';
 import StatusDefaultHome from './StatusDefaultHome/StatusDefaultHome';
 
 dayjs.extend(utc);
