--- src/pages/HomePage/ModalContents/index.ts
@@ -1,3 +1,0 @@
-export { default as ModalContentsAlert } from './ModalContentsAlert/ModalContentsAlert';
-export { default as ModalContentsCategory } from './ModalContentsCategory/ModalContentsCategory';
-export { default as ModalContentsFriends } from './ModalContentsFriends/ModalContentsFriends';
