+++ src/pages/HomePage/ModalContentsAlert/types/index.ts
