+++ src/pages/HomePage/ModalContentsFriends/FriendsList/components/FriendInfo.tsx
